refactor: consolidate database operations and improve service architecture
This commit implements a major refactoring to streamline database operations and improve the overall service architecture:

## Database Layer Consolidation
- **Enhanced database.ts**: Consolidated all database table operations into a centralized module
  - Added comprehensive CRUD operations for all entities (requests, equipment, songs, venues, items, members, assignees)
  - Implemented type-safe database operations with proper error handling
  - Added specialized query functions for complex joins and relationships
  - Renamed Val type to CellData for better clarity
  - Added PostgrestError import for improved error handling

## Service Layer Simplification
- **Removed duplicate services**: Eliminated board-service.ts and request-services.ts
  - These services contained duplicate database query logic
  - Functionality has been consolidated into database.ts and admin-service.ts
- **Streamlined admin-service.ts**: Reduced from complex query logic to focused business operations
  - Removed duplicate database queries that are now handled by database.ts
  - Maintained high-level business logic and API interfaces
  - Improved separation of concerns between data access and business logic

## Context Layer Updates
- **admin-context.tsx**: Updated to use consolidated database operations
  - Improved error handling and state management
  - Better integration with centralized database functions
- **board-context.tsx**: Simplified to use shared database operations
- **defaults-context.tsx**: Enhanced with consolidated data access patterns

## Hook Layer Improvements
- **use-cached-requests.ts**: Updated to work with consolidated database operations
  - Improved caching strategies
  - Better error handling and data synchronization

## Application Layer Enhancements
- **layout.tsx**: Added Progressive Web App (PWA) enhancements
  - Improved meta tags for better mobile experience
  - Enhanced theme configuration
- **request/page.tsx**: Optimized for better performance with consolidated data access

## Benefits of This Refactoring
- **Reduced Code Duplication**: Eliminated ~118 lines of duplicate database query code
- **Improved Maintainability**: Single source of truth for all database operations
- **Better Type Safety**: Enhanced TypeScript support with proper error handling
- **Cleaner Architecture**: Clear separation between data access, business logic, and presentation
- **Performance Improvements**: More efficient database queries and reduced redundancy
- **Easier Testing**: Centralized database operations are easier to mock and test

## Files Changed
- Modified: 8 files with enhanced functionality
- Deleted: 2 duplicate service files (board-service.ts, request-services.ts)
- Net result: -47 lines of code while adding more functionality

This refactoring maintains all existing functionality while providing a more robust, maintainable, and scalable foundation for future development.
